# api/openapi.yaml
openapi: 3.0.3
info:
  title: MailHive API
  description: |
    API multi-tenant d'envoi et de gestion d'e-mails avec file d'attente asynchrone,
    templates embarqués et chiffrement SMTP.
  version: 1.0.0
  contact:
    name: Franck Paszkowski
servers:
  - url: http://localhost:8080/api/v1
    description: Serveur local
...
security:
  - BearerAuth: []
paths:
...
  /auth/token:
    post:
      tags: [Auth]
      summary: Générer un token JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [api_key]
              properties:
                api_key:
                  type: string
                  description: Clé API du tenant ou de l'administrateur
      responses:
        "200":
          description: Token généré
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rafraîchir un token JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: Token JWT à rafraîchir
      responses:
        "200":
          description: Nouveau token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Health ────────────────────────────────────────
...
  /mails:
    post:
      tags: [Mails]
      summary: Créer et mettre en queue un ou plusieurs mails
      description: |
        Crée un mail et le met en file d'attente.

        **Mode groupé** (par défaut, `individuel: false`) : crée 1 seul mail avec tous les destinataires TO.
        La réponse est un objet `Mail`.

        **Mode individuel** (`individuel: true`) : crée N mails distincts (1 par destinataire TO).
        Chaque destinataire TO peut fournir un `template_data` spécifique qui sera fusionné avec
        le `template_data` partagé (les clés spécifiques écrasent les clés partagées).
        Les CC/BCC sont répliqués sur chaque mail. La réponse est un `CreateMailBatchResponse`.

        Le nombre total de destinataires (TO + CC + BCC) est limité par `max_destinataires` du tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMailRequest"
            examples:
              groupé:
                summary: Envoi groupé classique
                value:
                  template_id: "uuid-template"
                  template_data: { "entreprise": "Université de Lille" }
                  to:
                    - { "email": "alice@example.com", "name": "Alice" }
                    - { "email": "bob@example.com", "name": "Bob" }
              individuel:
                summary: Envoi individuel personnalisé
                value:
                  template_id: "uuid-template"
                  individuel: true
                  template_data:
                    { "entreprise": "Université de Lille", "annee": "2026" }
                  to:
                    - {
                        "email": "alice@example.com",
                        "name": "Alice",
                        "template_data": { "nom": "Alice", "code": "A123" },
                      }
                    - {
                        "email": "bob@example.com",
                        "name": "Bob",
                        "template_data": { "nom": "Bob", "code": "B456" },
                      }
                  cc:
                    - { "email": "manager@example.com" }
      responses:
        "202":
          description: |
            Mail(s) accepté(s) et mis en queue.
            - Mode groupé : retourne un `Mail`
            - Mode individuel : retourne un `CreateMailBatchResponse`
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/MailApiResponse"
                  - allOf:
                      - $ref: "#/components/schemas/ApiResponseBase"
                      - type: object
                        properties:
                          data:
                            $ref: "#/components/schemas/CreateMailBatchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
    get:
      tags: [Mails]
      summary: Lister les mails (pagination, filtre par statut)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/MailStatus"
      responses:
        "200":
          description: Liste paginée de mails
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailPaginatedResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /mails/stats:
    get:
      tags: [Mails]
      summary: Statistiques des mails du tenant
      responses:
        "200":
          description: Statistiques par statut
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/MailStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /mails/{id}:
    get:
      tags: [Mails]
      summary: Détail d'un mail avec destinataires
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Détail du mail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /mails/{id}/cancel:
    post:
      tags: [Mails]
      summary: Annuler un mail en attente
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Mail annulé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /mails/{id}/retry:
    post:
      tags: [Mails]
      summary: Relancer un mail en échec
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Mail relancé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  # ── Templates ─────────────────────────────────────
  /templates:
    post:
      tags: [Templates]
      summary: Créer un template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateRequest"
      responses:
        "201":
          description: Template créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      tags: [Templates]
      summary: Lister les templates
      responses:
        "200":
          description: Liste des templates
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Template"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /templates/{id}:
    get:
      tags: [Templates]
      summary: Détail d'un template
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Détail du template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Templates]
      summary: Modifier un template
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTemplateRequest"
      responses:
        "200":
          description: Template modifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Templates]
      summary: Supprimer un template
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Template supprimé
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /templates/{id}/preview:
    post:
      tags: [Templates]
      summary: Prévisualiser un template avec des données
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  type: object
                  additionalProperties:
                    type: string
                  description: Variables de substitution
      responses:
        "200":
          description: Rendu du template
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TemplatePreview"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── SMTP Configs ──────────────────────────────────
  /smtp-configs:
    post:
      tags: [SMTP Configs]
      summary: Créer une configuration SMTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSMTPConfigRequest"
      responses:
        "201":
          description: Config SMTP créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SMTPConfigApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      tags: [SMTP Configs]
      summary: Lister les configurations SMTP
      responses:
        "200":
          description: Liste des configs SMTP
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/SMTPConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /smtp-configs/{id}:
    get:
      tags: [SMTP Configs]
      summary: Détail d'une configuration SMTP
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Détail de la config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SMTPConfigApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [SMTP Configs]
      summary: Modifier une configuration SMTP
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSMTPConfigRequest"
      responses:
        "200":
          description: Config modifiée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SMTPConfigApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [SMTP Configs]
      summary: Supprimer une configuration SMTP
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Config supprimée
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /smtp-configs/{id}/test:
    post:
      tags: [SMTP Configs]
      summary: Tester la connexion SMTP
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Connexion SMTP réussie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Branding ──────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via POST /auth/token
...
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Non authentifié
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RateLimited:
      description: Limite de débit dépassée
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ApiResponseBase:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
    TokenResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data:
              type: object
              properties:
                token:
                  type: string

    # ── Health ──
...
    CreateMailBatchResponse:
      type: object
      description: Réponse lors d'un envoi individuel (N mails créés)
      properties:
        total:
          type: integer
          description: Nombre de mails créés
        mails:
          type: array
          items:
            $ref: "#/components/schemas/Mail"
...
    MailApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Mail"
...